Dont stop playing after closing Slider
